New clientlib ver

+++ KalturaClient/src/main/java/com/kaltura/client/types/FileAssetBaseFilter.java
@@ -61,7 +61,7 @@ public abstract class FileAssetBaseFilter extends RelatedFilter {
 		String statusIn();
 	}
 
-	private Integer idEqual;
+	private Long idEqual;
 	private String idIn;
 	private Integer partnerIdEqual;
 	private FileAssetObjectType fileAssetObjectTypeEqual;
@@ -75,10 +75,10 @@ public abstract class FileAssetBaseFilter extends RelatedFilter {
 	private String statusIn;
 
 	// idEqual:
-	public Integer getIdEqual(){
+	public Long getIdEqual(){
 		return this.idEqual;
 	}
-	public void setIdEqual(Integer idEqual){
+	public void setIdEqual(Long idEqual){
 		this.idEqual = idEqual;
 	}
 
@@ -229,7 +229,7 @@ public abstract class FileAssetBaseFilter extends RelatedFilter {
 		if(jsonObject == null) return;
 
 		// set members values:
-		idEqual = GsonParser.parseInt(jsonObject.get("idEqual"));
+		idEqual = GsonParser.parseLong(jsonObject.get("idEqual"));
 		idIn = GsonParser.parseString(jsonObject.get("idIn"));
 		partnerIdEqual = GsonParser.parseInt(jsonObject.get("partnerIdEqual"));
 		fileAssetObjectTypeEqual = FileAssetObjectType.get(GsonParser.parseString(jsonObject.get("fileAssetObjectTypeEqual")));
@@ -282,7 +282,7 @@ public abstract class FileAssetBaseFilter extends RelatedFilter {
 
     public FileAssetBaseFilter(Parcel in) {
         super(in);
-        this.idEqual = (Integer)in.readValue(Integer.class.getClassLoader());
+        this.idEqual = (Long)in.readValue(Long.class.getClassLoader());
         this.idIn = in.readString();
         this.partnerIdEqual = (Integer)in.readValue(Integer.class.getClassLoader());
         int tmpFileAssetObjectTypeEqual = in.readInt();
