+++ KalturaClient/src/main/java/com/kaltura/client/types/ResponseProfileBaseFilter.java
@@ -58,7 +58,7 @@ public abstract class ResponseProfileBaseFilter extends Filter {
 		String statusIn();
 	}
 
-	private Integer idEqual;
+	private Long idEqual;
 	private String idIn;
 	private String systemNameEqual;
 	private String systemNameIn;
@@ -70,10 +70,10 @@ public abstract class ResponseProfileBaseFilter extends Filter {
 	private String statusIn;
 
 	// idEqual:
-	public Integer getIdEqual(){
+	public Long getIdEqual(){
 		return this.idEqual;
 	}
-	public void setIdEqual(Integer idEqual){
+	public void setIdEqual(Long idEqual){
 		this.idEqual = idEqual;
 	}
 
@@ -200,7 +200,7 @@ public abstract class ResponseProfileBaseFilter extends Filter {
 		if(jsonObject == null) return;
 
 		// set members values:
-		idEqual = GsonParser.parseInt(jsonObject.get("idEqual"));
+		idEqual = GsonParser.parseLong(jsonObject.get("idEqual"));
 		idIn = GsonParser.parseString(jsonObject.get("idIn"));
 		systemNameEqual = GsonParser.parseString(jsonObject.get("systemNameEqual"));
 		systemNameIn = GsonParser.parseString(jsonObject.get("systemNameIn"));
@@ -247,7 +247,7 @@ public abstract class ResponseProfileBaseFilter extends Filter {
 
     public ResponseProfileBaseFilter(Parcel in) {
         super(in);
-        this.idEqual = (Integer)in.readValue(Integer.class.getClassLoader());
+        this.idEqual = (Long)in.readValue(Long.class.getClassLoader());
         this.idIn = in.readString();
         this.systemNameEqual = in.readString();
         this.systemNameIn = in.readString();
